requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
passlib==1.7.4
aiofiles==23.2.1